# packages/sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "c-plane API",
    "description": "Control plane API for managing projects, environments, containers, and databases",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/organization/{organization_id}/registry": {
      "get": {
        "tags": [
          "registry"
        ],
        "operationId": "get_registry",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Managed Registry configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedRegistryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Managed Registry is not activated"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-cplane-required-scope": "registry:read"
      },
      "put": {
        "tags": [
          "registry"
        ],
        "operationId": "activate_registry",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateManagedRegistryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Managed Registry already active",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedRegistryResponse"
                }
              }
            }
          },
          "201": {
            "description": "Managed Registry activated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ManagedRegistryResponse"
                }
              }
            }
          },
          "404": {
            "description": "Organization or active region not found"
          },
          "409": {
            "description": "Active region has no S3 provider"
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "x-cplane-required-scope": "registry:create"
      }
    },
...
    "/api/registry/token": {
      "get": {
        "tags": [
          "registry"
        ],
        "operationId": "issue_token",
        "parameters": [
          {
            "name": "service",
            "in": "query",
            "description": "Distribution service name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "description": "Requested repository scopes",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Short-lived Distribution access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid registry credentials"
          },
          "503": {
            "description": "Registry is unavailable during maintenance",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "registryBasic": []
          }
        ],
        "x-cplane-required-scope": [
          "registry:pull",
          "registry:push"
        ]
      }
    },
...
    "/internal/organizations/{organization_id}/registry": {
      "get": {
        "tags": [
          "internal"
        ],
        "operationId": "resolve_registry",
        "parameters": [
          {
            "name": "organization_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolvedManagedRegistry"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "serviceToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActivateManagedRegistryRequest": {
        "type": "object",
        "required": [
          "region_id"
        ],
        "properties": {
          "region_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "details": {},
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ManagedRegistryResponse": {
        "type": "object",
        "required": [
          "organization_id",
          "region_id",
          "status",
          "storage_revision",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "region_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "storage_revision": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "RegistryTokenResponse": {
        "type": "object",
        "required": [
          "token",
          "access_token",
          "expires_in",
          "issued_at"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "expires_in": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "issued_at": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
...
      "ResolvedManagedRegistry": {
        "type": "object",
        "required": [
          "organization_id",
          "organization_slug",
          "storage_revision",
          "status",
          "access_key_id",
          "bucket_name",
          "storage_endpoint_url"
        ],
        "properties": {
          "access_key_id": {
            "type": "string"
          },
          "bucket_name": {
            "type": "string"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid"
          },
          "organization_slug": {
            "type": "string"
          },
          "repository_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "repository_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          },
          "storage_endpoint_url": {
            "type": "string"
          },
          "storage_revision": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key with the operation's required scope."
      },
      "registryBasic": {
        "type": "http",
        "scheme": "basic"
      },
      "serviceToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-cplane-token",
        "description": "Internal service token."
      }
    }
  },
...
}